requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
et_xmlfile==2.0.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
idna==3.10
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2==2.9.10
pyarrow==19.0.1
pyogrio==0.10.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
shapely==2.0.7
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
